chore: rename dir to camelCase
