+++ src/timeMachine/components/fluxFunctionsToolbar/FunctionCategory.tsx
@@ -3,7 +3,7 @@ import React, {SFC} from 'react'
 
 // Components
 import ToolbarFunction from 'src/flows/shared/FilterList/InjectionOption'
-import FunctionTooltipContent from 'src/flows/pipes/RawFluxEditor/FunctionsList/per-function/FunctionToolTipContent'
+import FunctionTooltipContent from 'src/flows/pipes/RawFluxEditor/FunctionsList/perFunction/FunctionToolTipContent'
 
 // Types
 import {FluxToolbarFunction} from 'src/types/shared'
